+++ src/main/java/com/github/camilormz/finalreality/model/character/AbstractCharacter.java
@@ -1,9 +1,12 @@
 package com.github.camilormz.finalreality.model.character;
 
+import java.beans.PropertyChangeListener;
 import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
+
+import java.beans.PropertyChangeSupport;
 
 import com.github.camilormz.finalreality.model.character.player.IPlayerCharacter;
 import org.jetbrains.annotations.NotNull;
@@ -22,6 +25,14 @@ public abstract class AbstractCharacter implements ICharacter {
   private int healthPoints;
   private final int defense;
   private ScheduledExecutorService scheduledExecutor;
+
+  private final PropertyChangeSupport addToQueueEvent = new PropertyChangeSupport(this);
+  protected final PropertyChangeSupport knockOutEvent = new PropertyChangeSupport(this);
+
+  private final String addToQueueEventName = String.format(
+                                             "Character enqueued: %d", this.hashCode());
+  protected final String knockOutEventName = String.format(
+                                             "Character knocked out: %d", this.hashCode());
 
   /**
    * Creates a new (abstract) character.
@@ -63,6 +74,7 @@ public abstract class AbstractCharacter implements ICharacter {
   private void addToQueue() {
     turnsQueue.add(this);
     scheduledExecutor.shutdown();
+    addToQueueEvent.firePropertyChange(addToQueueEventName, null, this);
   }
 
   /**
@@ -132,4 +144,17 @@ public abstract class AbstractCharacter implements ICharacter {
   @Override
   public abstract void beAttackedByEnemy(Enemy enemy);
 
+  /**
+   * Adds a handler for the enqueuing event
+   */
+  public void addEnqueuingListener(PropertyChangeListener handler) {
+    addToQueueEvent.addPropertyChangeListener(handler);
+  }
+
+  /**
+   * Adds a handler for knock out events
+   */
+  public void addKnockOutListener(PropertyChangeListener handler) {
+    knockOutEvent.addPropertyChangeListener(handler);
+  }
 }